Test: Run focus navigation in Browser Mode
