+++ packages/components/src/modal/test/index.jsdom.test.tsx
@@ -7,10 +7,8 @@ import {
 	useEnableWpCompatOverlaySlot,
 } from '@wordpress/ui';
 import Modal from '../';
-import type { ModalProps } from '../types';
 globalThis.wpVitest.mockCSSSupports();
 globalThis.wpVitest.mockMatchMedia();
-globalThis.wpVitest.mockVisibleElements();
 
 const noop = () => {};
 
@@ -320,106 +318,6 @@ describe( 'Modal', () => {
 		expect(
 			screen.getByText( 'A sweet button', { selector: 'button' } )
 		).toBeInTheDocument();
-	} );
-
-	describe( 'Focus handling', () => {
-		const FocusMountDemo = ( {
-			focusOnMount,
-		}: Pick< ModalProps, 'focusOnMount' > ) => {
-			const [ isShown, setIsShown ] = useState( false );
-			return (
-				<>
-					<button onClick={ () => setIsShown( true ) }>
-						Toggle Modal
-					</button>
-					{ isShown && (
-						<Modal
-							focusOnMount={ focusOnMount }
-							onRequestClose={ () => setIsShown( false ) }
-						>
-							<p>Modal content</p>
-							<a href="https://wordpress.org">
-								First Focusable Content Element
-							</a>
-
-							<a href="https://wordpress.org">
-								Another Focusable Content Element
-							</a>
-						</Modal>
-					) }
-				</>
-			);
-		};
-
-		it( 'should focus the Modal dialog by default when `focusOnMount` prop is not provided', async () => {
-			const user = userEvent.setup();
-
-			render( <FocusMountDemo /> );
-
-			const opener = screen.getByRole( 'button', {
-				name: 'Toggle Modal',
-			} );
-
-			await user.click( opener );
-
-			expect( screen.getByRole( 'dialog' ) ).toHaveFocus();
-		} );
-
-		it( 'should focus the Modal dialog when `true` passed as value for `focusOnMount` prop', async () => {
-			const user = userEvent.setup();
-
-			render( <FocusMountDemo focusOnMount /> );
-
-			const opener = screen.getByRole( 'button', {
-				name: 'Toggle Modal',
-			} );
-
-			await user.click( opener );
-
-			expect( screen.getByRole( 'dialog' ) ).toHaveFocus();
-		} );
-
-		it( 'should focus the first focusable element in the contents (if found) when `firstContentElement` passed as value for `focusOnMount` prop', async () => {
-			const user = userEvent.setup();
-
-			render( <FocusMountDemo focusOnMount="firstContentElement" /> );
-
-			const opener = screen.getByRole( 'button' );
-
-			await user.click( opener );
-
-			expect(
-				screen.getByText( 'First Focusable Content Element' )
-			).toHaveFocus();
-		} );
-
-		it( 'should focus the first element anywhere within the Modal when `firstElement` passed as value for `focusOnMount` prop', async () => {
-			const user = userEvent.setup();
-
-			render( <FocusMountDemo focusOnMount="firstElement" /> );
-
-			const opener = screen.getByRole( 'button' );
-
-			await user.click( opener );
-
-			expect(
-				screen.getByRole( 'button', { name: 'Close' } )
-			).toHaveFocus();
-		} );
-
-		it( 'should not move focus when `false` passed as value for `focusOnMount` prop', async () => {
-			const user = userEvent.setup();
-
-			render( <FocusMountDemo focusOnMount={ false } /> );
-
-			const opener = screen.getByRole( 'button', {
-				name: 'Toggle Modal',
-			} );
-
-			await user.click( opener );
-
-			expect( opener ).toHaveFocus();
-		} );
 	} );
 
 	describe( 'Body class name', () => {
